chore(backend): move shared edit_distance into ids.py
Delete the new text.py module so memory_ingestion stays within its
grandfathered source-file count, and point pipeline/verify_output at ids.
