--- backend/utils/memory_ingestion/text.py
@@ -1,21 +1,0 @@
-"""Small text helpers shared across memory ingestion stages."""
-
-from __future__ import annotations
-
-
-def edit_distance(a: str, b: str) -> int:
-    """Levenshtein edit distance between two strings."""
-    if len(a) < len(b):
-        return edit_distance(b, a)
-    if len(b) == 0:
-        return len(a)
-    prev_row = list(range(len(b) + 1))
-    for i, ca in enumerate(a):
-        curr_row = [i + 1]
-        for j, cb in enumerate(b):
-            insertions = prev_row[j + 1] + 1
-            deletions = curr_row[j] + 1
-            substitutions = prev_row[j] + (ca != cb)
-            curr_row.append(min(insertions, deletions, substitutions))
-        prev_row = curr_row
-    return prev_row[-1]
